+++ api/python/src/PE/objects/pySection.cpp
@@ -16,22 +16,64 @@
 #include "PE/pyPE.hpp"
 
 #include "LIEF/PE/Section.hpp"
+#include "enums_wrapper.hpp"
 
 #include <string>
 #include <sstream>
 #include <nanobind/stl/string.h>
 #include <nanobind/stl/set.h>
 
+#define PY_ENUM(x) to_string(x), x
+
 namespace LIEF::PE::py {
 
 template<>
 void create<Section>(nb::module_& m) {
-  nb::class_<Section, LIEF::Section>(m, "Section",
+  nb::class_<Section, LIEF::Section> sec(m, "Section",
       R"delim(
       Class which represents a PE section.
 
       It extends the base class :class:`lief.Section`
-      )delim"_doc)
+      )delim"_doc);
+
+  enum_<Section::CHARACTERISTICS>(sec, "CHARACTERISTICS", nb::is_arithmetic())
+    .value(PY_ENUM(Section::CHARACTERISTICS::TYPE_NO_PAD))
+    .value(PY_ENUM(Section::CHARACTERISTICS::CNT_CODE))
+    .value(PY_ENUM(Section::CHARACTERISTICS::CNT_INITIALIZED_DATA))
+    .value(PY_ENUM(Section::CHARACTERISTICS::CNT_UNINITIALIZED_DATA))
+    .value(PY_ENUM(Section::CHARACTERISTICS::LNK_OTHER))
+    .value(PY_ENUM(Section::CHARACTERISTICS::LNK_INFO))
+    .value(PY_ENUM(Section::CHARACTERISTICS::LNK_REMOVE))
+    .value(PY_ENUM(Section::CHARACTERISTICS::LNK_COMDAT))
+    .value(PY_ENUM(Section::CHARACTERISTICS::GPREL))
+    .value(PY_ENUM(Section::CHARACTERISTICS::MEM_PURGEABLE))
+    .value(PY_ENUM(Section::CHARACTERISTICS::MEM_16BIT))
+    .value(PY_ENUM(Section::CHARACTERISTICS::MEM_LOCKED))
+    .value(PY_ENUM(Section::CHARACTERISTICS::MEM_PRELOAD))
+    .value(PY_ENUM(Section::CHARACTERISTICS::ALIGN_1BYTES))
+    .value(PY_ENUM(Section::CHARACTERISTICS::ALIGN_2BYTES))
+    .value(PY_ENUM(Section::CHARACTERISTICS::ALIGN_4BYTES))
+    .value(PY_ENUM(Section::CHARACTERISTICS::ALIGN_8BYTES))
+    .value(PY_ENUM(Section::CHARACTERISTICS::ALIGN_16BYTES))
+    .value(PY_ENUM(Section::CHARACTERISTICS::ALIGN_32BYTES))
+    .value(PY_ENUM(Section::CHARACTERISTICS::ALIGN_64BYTES))
+    .value(PY_ENUM(Section::CHARACTERISTICS::ALIGN_128BYTES))
+    .value(PY_ENUM(Section::CHARACTERISTICS::ALIGN_256BYTES))
+    .value(PY_ENUM(Section::CHARACTERISTICS::ALIGN_512BYTES))
+    .value(PY_ENUM(Section::CHARACTERISTICS::ALIGN_1024BYTES))
+    .value(PY_ENUM(Section::CHARACTERISTICS::ALIGN_2048BYTES))
+    .value(PY_ENUM(Section::CHARACTERISTICS::ALIGN_4096BYTES))
+    .value(PY_ENUM(Section::CHARACTERISTICS::ALIGN_8192BYTES))
+    .value(PY_ENUM(Section::CHARACTERISTICS::LNK_NRELOC_OVFL))
+    .value(PY_ENUM(Section::CHARACTERISTICS::MEM_DISCARDABLE))
+    .value(PY_ENUM(Section::CHARACTERISTICS::MEM_NOT_CACHED))
+    .value(PY_ENUM(Section::CHARACTERISTICS::MEM_NOT_PAGED))
+    .value(PY_ENUM(Section::CHARACTERISTICS::MEM_SHARED))
+    .value(PY_ENUM(Section::CHARACTERISTICS::MEM_EXECUTE))
+    .value(PY_ENUM(Section::CHARACTERISTICS::MEM_READ))
+    .value(PY_ENUM(Section::CHARACTERISTICS::MEM_WRITE));
+
+  sec
     .def(nb::init<>())
     .def(nb::init<const std::vector<uint8_t>&, const std::string&, uint32_t>(),
         "Constructor from "
@@ -51,7 +93,8 @@ void create<Section>(nb::module_& m) {
         R"delim(
         The total size of the section when loaded into memory.
 
-        If this value is greater than :attr:`~lief.PE.Section.sizeof_raw_data`, the section is zero-padded.
+        If this value is greater than :attr:`~lief.PE.Section.sizeof_raw_data`,
+        the section is zero-padded.
         )delim"_doc)
 
     .def_prop_rw("sizeof_raw_data",
@@ -68,8 +111,9 @@ void create<Section>(nb::module_& m) {
         nb::overload_cast<>(&Section::pointerto_relocation, nb::const_),
         nb::overload_cast<uint32_t>(&Section::pointerto_relocation),
         R"delim(
-        The file pointer to the beginning of the COFF relocation entries for the section. This is set to zero for
-        executable images or if there are no relocations.
+        The file pointer to the beginning of the COFF relocation entries for
+        the section. This is set to zero for executable images or if there are
+        no relocations.
 
         For modern PE binaries, this value is usually set to 0 as the relocations are managed by
         :class:`~lief.PE.Relocation`.
@@ -80,8 +124,9 @@ void create<Section>(nb::module_& m) {
         nb::overload_cast<uint32_t>(&Section::pointerto_line_numbers),
         R"delim(
         The file pointer to the beginning of line-number entries for the section.
-        This is set to zero if there are no COFF line numbers. This value should be zero for an image because COFF
-        debugging information is deprecated and modern debug information relies on the PDB files.
+        This is set to zero if there are no COFF line numbers. This value should
+        be zero for an image because COFF debugging information is deprecated
+        and modern debug information relies on the PDB files.
         )delim"_doc)
 
     .def_prop_rw("numberof_relocations",
@@ -108,7 +153,7 @@ void create<Section>(nb::module_& m) {
     .def_prop_rw("characteristics",
         nb::overload_cast<>(&Section::characteristics, nb::const_),
         nb::overload_cast<uint32_t>(&Section::characteristics),
-        "The " RST_CLASS_REF(lief.PE.SECTION_CHARACTERISTICS) "  that describe the characteristics of the section"_doc)
+        "The " RST_CLASS_REF(lief.PE.Section.CHARACTERISTICS) "  that describe the characteristics of the section"_doc)
 
     .def_prop_ro("characteristics_lists",
         &Section::characteristics_list,
@@ -116,7 +161,7 @@ void create<Section>(nb::module_& m) {
 
     .def("has_characteristic",
         &Section::has_characteristic,
-        "``True`` if the section has the given " RST_CLASS_REF(lief.PE.SECTION_CHARACTERISTICS) ""_doc,
+        "``True`` if the section has the given " RST_CLASS_REF(lief.PE.Section.CHARACTERISTICS) ""_doc,
         "characteristic"_a)
 
     .def_prop_ro("padding",