+++ src/PE/Binary.cpp
@@ -494,17 +494,17 @@ Section* Binary::add_section(const Section& section, PE_SECTION_TYPES type) {
   }
 
   if (new_section->is_type(PE_SECTION_TYPES::TEXT)) {
-    new_section->add_characteristic(SECTION_CHARACTERISTICS::IMAGE_SCN_CNT_CODE);
-    new_section->add_characteristic(SECTION_CHARACTERISTICS::IMAGE_SCN_MEM_EXECUTE);
-    new_section->add_characteristic(SECTION_CHARACTERISTICS::IMAGE_SCN_MEM_READ);
+    new_section->add_characteristic(Section::CHARACTERISTICS::CNT_CODE);
+    new_section->add_characteristic(Section::CHARACTERISTICS::MEM_EXECUTE);
+    new_section->add_characteristic(Section::CHARACTERISTICS::MEM_READ);
     optional_header().baseof_code(static_cast<uint32_t>(new_section->virtual_address()));
     optional_header().sizeof_code(new_section->sizeof_raw_data());
   }
 
   if (new_section->is_type(PE_SECTION_TYPES::DATA)) {
-    new_section->add_characteristic(SECTION_CHARACTERISTICS::IMAGE_SCN_CNT_INITIALIZED_DATA);
-    new_section->add_characteristic(SECTION_CHARACTERISTICS::IMAGE_SCN_MEM_READ);
-    new_section->add_characteristic(SECTION_CHARACTERISTICS::IMAGE_SCN_MEM_WRITE);
+    new_section->add_characteristic(Section::CHARACTERISTICS::CNT_INITIALIZED_DATA);
+    new_section->add_characteristic(Section::CHARACTERISTICS::MEM_READ);
+    new_section->add_characteristic(Section::CHARACTERISTICS::MEM_WRITE);
 
     if (this->type() == PE_TYPE::PE32) {
       optional_header().baseof_data(static_cast<uint32_t>(new_section->virtual_address()));
@@ -514,9 +514,9 @@ Section* Binary::add_section(const Section& section, PE_SECTION_TYPES type) {
 
 
   if (type == PE_SECTION_TYPES::IMPORT) {
-    new_section->add_characteristic(SECTION_CHARACTERISTICS::IMAGE_SCN_MEM_READ);
-    new_section->add_characteristic(SECTION_CHARACTERISTICS::IMAGE_SCN_MEM_EXECUTE);
-    new_section->add_characteristic(SECTION_CHARACTERISTICS::IMAGE_SCN_MEM_WRITE);
+    new_section->add_characteristic(Section::CHARACTERISTICS::MEM_READ);
+    new_section->add_characteristic(Section::CHARACTERISTICS::MEM_EXECUTE);
+    new_section->add_characteristic(Section::CHARACTERISTICS::MEM_WRITE);
 
     data_directory(DataDirectory::TYPES::IMPORT_TABLE)->RVA(new_section->virtual_address());
     data_directory(DataDirectory::TYPES::IMPORT_TABLE)->size(new_section->sizeof_raw_data());